mac_agent/requirements.txt
httpx==0.28.1
mss==10.1.0
Pillow==11.3.0
pyautogui==0.9.54
python-dotenv==1.1.1
websockets==15.0.1
